Fixed unnecessary state in NavBar
